LayoutTests/fast/selectors/querySelector-multiple-selectors.html

<!doctype html>
<html id=htmlDocument>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<div style="display:none">
<t0>
    <t0></t0>
</t0>
<t1>
    <t0></t0>
</t1>


<t3></t3>

<t4>
    <t5 id="t5"></t5>
</t4>
<t6>
    <ng>
        <t7 class="ok"></t7>
    </ng>
</t6>

<t8></t8>
<t9></t9>
<t9></t9>
<t10>
    <t11 class="ng"></t11>
</t10>

<ok class="t12"></ok>
<ok class="t13 t14 t15"></ok>

<ok class="t16"></ok>
<ok class="t17">
    <t18></t18>
    <t18></t18>
    <t18></t18>
    <t18></t18>
    <t18></t18>
</ok>
<ok class="t17">
    <t18></t18>
    <t18></t18>
    <t18></t18>
    <t18></t18>
    <t18></t18>
</ok>

</div>
</body>
<script>
description('Test cases when the multiple selectors are provided');

shouldBe('document.body.querySelectorAll("t0, t1").length', '4');
shouldBe('document.body.querySelectorAll("#t2, t3").length', '1');
shouldBe('document.body.querySelectorAll("t4> #t5, t6 t7.ok").length', '2');
shouldBe('document.body.querySelectorAll("t8 + t9, t10 t11.ok").length', '1');
shouldBe('document.body.querySelectorAll("t8 + t9, t10 t11.ok").length', '1');
shouldBe('document.body.querySelectorAll(".t12, .t13, .t14, .t15").length', '2');

debug('CSS JIT compile may fail, fallback to the slow path.')
shouldBe('document.body.querySelectorAll(".t16, .t17>t18:nth-last-child(1)").length', '3');
</script>
<script src="../../resources/js-test-post.js"></script>
</html>
